[Oztechan/CCC#2605] Move PremiumView to PremiumRootView
